Feat(SecureMe): Support PicoW
Implements support for the PicoW series of microcontrollers.

This update adds conditional support for both Pico and PicoW microcontrollers for features such as onboard LED.
It additionally implements a captive portal which enables easy connection of the firmware to a wi-fi network.
The captive portal saves network configuration so that automatic connection can be achieved during initialization. If the network is unavailable, the captive portal access point will be initialized to enable connection to a new network.