Feat(Config): Configurable sensor cooldown
Implements a configurable sensor cooldown time.
The sensor cooldown time is the time the SecureMe system will wait after detection before detecting again.
The cooldown is applied per sensor per detection.
By default the cooldown is set to 10 seconds which was the previous behaviour before making the cooldown configurable.
The cooldown time can be adjusted in the detection settings page of the SecureMe web management portal.